front-end: initialize setup

+++ backend/controllers/product.controller.js
@@ -42,7 +42,6 @@ export const updateProduct = async (req, res) =>{
     const product = req.body;
 
     if(!mongoose.Types.ObjectId.isValid(id)){
-
         return res.status(404).json({success: false, message:'Invalid object ID.'})
     };
 
@@ -61,13 +60,18 @@ export const updateProduct = async (req, res) =>{
 export const deleteProduct = async (req, res) => {
 
     const {id} = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success: false, message:'Invalid object ID.'})
+    };
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product deleted."});
     }
     catch(err){
         console.log('error in deleting product:', error.message);
-        res.status(501).json({success: false, message: 'Product not found.'});
+        res.status(500).json({success: false, message: 'Server error.'});
     }
 };
